fastapi-microservice/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
backoff==2.2.1
certifi==2024.8.30
click==8.1.7
colorama==0.4.6
dishka==1.3.0
dnspython==2.6.1
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.114.2
frozenlist==1.4.1
h11==0.14.0
idna==3.9
multidict==6.1.0
orjson==3.10.7
passlib==1.7.4
pyasn1==0.6.1
pydantic==2.9.1
pydantic-settings==2.5.2
pydantic_core==2.23.3
python-dotenv==1.0.1
python-jose==3.3.0
redis==5.0.8
rsa==4.9
sentry-sdk==2.14.0
six==1.16.0
sniffio==1.3.1
starlette==0.38.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.30.6
yarl==1.11.1
